Added exp-lbrs-replace2.ulp
Here is the description by the author  at the moment.


I modified the export project ULP to allow exporting all parts into a single project specific library. BUT, the magic part is that it also generates a script you can run afterward to replace all parts in the schematic so they reference the newly created project specific library. I\'ve also updated the ULP so it now shows a list of all libraries used in a schematic and it allows selecting which library parts to export/replace. Now you can easily export and edit any of the parts in any of your designs within a single project specific library on the fly!